requirements.txt
alembic==1.13.1
aniso8601==9.0.1
Babel==2.15.0
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Admin==1.6.1
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-Uploads==0.2.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
passlib==1.7.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.30
typing_extensions==4.12.2
Werkzeug==3.0.3
WTForms==3.1.2
